Refactor: 기업인증 관련 파일 정리 및 폴더 구조 개선
