--- src/components/ReceiptOCR.module.css
@@ -1,81 +1,0 @@
-.container {
-    max-width: 800px;
-    margin: 0 auto;
-    padding: 20px;
-  }
-  
-  .form {
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    margin-top: 20px;
-  }
-  
-  .fileInput {
-    padding: 10px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    width: 100%;
-  }
-  
-  .preview {
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    padding: 10px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: #f9f9f9;
-    min-height: 200px;
-  }
-  
-  .preview img {
-    max-width: 100%;
-    max-height: 300px;
-    object-fit: contain;
-  }
-  
-  .submitButton {
-    padding: 10px 20px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 16px;
-  }
-  
-  .submitButton:disabled {
-    background-color: #ccc;
-    cursor: not-allowed;
-  }
-  
-  .submitButton:hover:not(:disabled) {
-    background-color: #0056b3;
-  }
-  
-  .error {
-    color: #dc3545;
-    padding: 10px;
-    border: 1px solid #dc3545;
-    border-radius: 4px;
-    margin-top: 20px;
-    background-color: #f8d7da;
-  }
-  
-  .result {
-    margin-top: 20px;
-    padding: 20px;
-    border: 1px solid #28a745;
-    border-radius: 4px;
-    background-color: #f8f9fa;
-  }
-  
-  .result h3 {
-    margin-top: 0;
-    color: #28a745;
-  }
-  
-  .result p {
-    margin: 10px 0;
-  }
