+++ src/styles/mypage/BusinessRegistrationOCR.module.css
